requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
dash
folium
textblob
wordcloud
python-dotenv 